fix key passthrough in trace-util
