+++ lib/api/index.js
@@ -9,6 +9,8 @@ const path = require("path"),
   LibhoneyImpl = require("./libhoney"),
   MockImpl = require("./mock");
 
+const headers = propagation.honeycomb;
+
 let apiImpl;
 
 function mapFieldsToApp(map) {
@@ -49,12 +51,12 @@ module.exports = {
     return tracker.getTracked();
   },
 
-  TRACE_HTTP_HEADER: propagation.TRACE_HTTP_HEADER,
-  AMAZON_TRACE_HTTP_HEADER: propagation.AMAZON_TRACE_HTTP_HEADER,
-  REQUEST_ID_HTTP_HEADER: propagation.REQUEST_ID_HTTP_HEADER,
-  OTEL_HTTP_HEADER: propagation.OTEL_HTTP_HEADER,
-  marshalTraceContext: propagation.marshalTraceContext,
-  unmarshalTraceContext: propagation.unmarshalTraceContext,
+  TRACE_HTTP_HEADER: headers.TRACE_HTTP_HEADER,
+  AMAZON_TRACE_HTTP_HEADER: headers.AMAZON_TRACE_HTTP_HEADER,
+  REQUEST_ID_HTTP_HEADER: headers.REQUEST_ID_HTTP_HEADER,
+  OTEL_HTTP_HEADER: headers.OTEL_HTTP_HEADER,
+  marshalTraceContext: headers.marshalTraceContext,
+  unmarshalTraceContext: headers.unmarshalTraceContext,
 
   startTrace(metadataContext, withTraceId, withParentSpanId, withDataset) {
     return apiImpl.startTrace(metadataContext, withTraceId, withParentSpanId, withDataset);
