+++ lib/instrumentation/express.js
@@ -83,9 +83,8 @@ const getMagicMiddleware = ({ userContext, traceIdSource, parentIdSource, packag
   res,
   next
 ) => {
-  console.log(req.headers);
-  console.log(traceIdSource);
   let traceContext = traceUtil.getTraceContext(traceIdSource, req);
+  console.log(traceContext);
   let parentTraceId = traceUtil.getParentSourceId(parentIdSource, req);
   if (parentTraceId) {
     traceContext.parentSpanId = parentTraceId;
@@ -121,8 +120,6 @@ const getMagicMiddleware = ({ userContext, traceIdSource, parentIdSource, packag
     traceContext.parentSpanId,
     traceContext.dataset
   );
-  console.log(traceContext);
-  console.log(parentTraceId);
 
   if (traceContext.customContext) {
     api.addTraceContext(traceContext.customContext);
