+++ lib/instrumentation/trace-util.js
@@ -4,8 +4,8 @@ const { parseOtelTrace } = require("../propagation/w3c_context");
 
 // returns the header name/value for the first header with a value in the request.
 const getValueFromHeaders = (requestHeaders, headers) => {
-  console.log(requestHeaders, headers);
   let value, header;
+  console.log(headers);
 
   for (const h of headers) {
     let headerValue = requestHeaders[h.toLowerCase()];
@@ -24,6 +24,7 @@ const getValueFromHeaders = (requestHeaders, headers) => {
 };
 
 exports.getTraceContext = (traceIdSource, req) => {
+  //console.log(req);
   if (typeof traceIdSource === "undefined" || typeof traceIdSource === "string") {
     let headers =
       typeof traceIdSource === "undefined"
